Move nonvoting/internal/s11n to internal/s11n
