+++ internal/s11n/descriptor_test.go
