+++ internal/s11n/document.go
