+++ internal/s11n/document_test.go
