+++ nonvoting/server/state.go
@@ -26,7 +26,7 @@ import (
 	"time"
 
 	bolt "github.com/coreos/bbolt"
-	"github.com/katzenpost/authority/nonvoting/internal/s11n"
+	"github.com/katzenpost/authority/internal/s11n"
 	"github.com/katzenpost/core/crypto/eddsa"
 	"github.com/katzenpost/core/crypto/rand"
 	"github.com/katzenpost/core/epochtime"
